% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot_group}
\alias{plot_group}
\title{Title}
\usage{
plot_group(
  data,
  group,
  feature = "SID_Score",
  comparisons = NULL,
  map_signif_level = FALSE
)
}
\arguments{
\item{data}{}

\item{group}{}

\item{comparisons}{}

\item{map_signif_level}{}
}
\description{
Title
}
