% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPro.R
\name{scale_data}
\alias{scale_data}
\title{Title}
\usage{
scale_data(data, denoising = TRUE, threads = 1, savetmp = FALSE)
}
\arguments{
\item{data}{}

\item{denoising}{}

\item{threads}{}

\item{savetmp}{}
}
\description{
Title
}
