% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{SenCID}
\alias{SenCID}
\title{Title}
\usage{
SenCID(
  data,
  sidnums = c(1, 2, 3, 4, 5, 6),
  binarize = FALSE,
  denoising,
  threads
)
}
\arguments{
\item{sidnums}{}

\item{denoising}{}

\item{threads}{}

\item{adata}{}
}
\description{
Title
}
