% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{sene_heatmap}
\alias{sene_heatmap}
\title{Title}
\usage{
sene_heatmap(CDS_obj, num = 20, num_clusters = 3, cores = 3)
}
\arguments{
\item{CDS_obj}{}

\item{num}{}

\item{num_clusters}{}

\item{cores}{}
}
\description{
Title
}
