% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{sene_traject}
\alias{sene_traject}
\title{Title}
\usage{
sene_traject(seurat_obj, color = "SID_Score", cores = 3)
}
\arguments{
\item{seurat_obj}{}

\item{color}{}

\item{cores}{}
}
\description{
Title
}
